+++ src/lib/utils/git.ts
@@ -1,16 +1,20 @@
-import * as childProcess from "child_process";
-import {promisify} from "util";
+import {AbsPosixPath} from "../types";
+import {ExecFileOptions, SpawnedProcess, SpawnResult} from "./node-async";
 
-type ExecFileAsync = (file: string, args?: string[], options?: any) => Promise<any>;
-const execFileAsync: ExecFileAsync = <any> promisify(childProcess.execFile);
-
-export async function exec(cmd: string, args: string[] = [], options?: any): Promise<Buffer> {
-  return execFileAsync("git", [cmd, ...args], options);
+export async function execGit(cmd: string, args: string[] = [], options?: ExecFileOptions): Promise<SpawnResult> {
+  return new SpawnedProcess(
+    "git",
+    [cmd, ...args],
+    {
+      cwd: options !== undefined ? options.cwd : undefined,
+      stdio: "inherit",
+    },
+  ).toPromise();
 }
 
 export async function assertCleanBranch(allowedBranches: string[]): Promise<void> {
   let stdout: Buffer;
-  stdout = await exec("symbolic-ref", ["HEAD"]);
+  stdout = (await execGit("symbolic-ref", ["HEAD"])).stdout;
   let onAllowedBranch: boolean = false;
   for (const branch of allowedBranches) {
     if (stdout.toString("utf8").trim() === `refs/heads/${branch}`) {
@@ -20,13 +24,82 @@ export async function assertCleanBranch(allowedBranches: string[]): Promise<void
   if (!onAllowedBranch) {
     throw new Error(`HEAD must be on one of the branches: ${JSON.stringify(allowedBranches)}`);
   }
-  stdout = await exec("status", ["--porcelain"]);
+  stdout = (await execGit("status", ["--porcelain"])).stdout;
   if (stdout.toString("utf8").trim().length > 0) {
     throw new Error("Working copy is dirty");
   }
 }
 
+/**
+ * Get the hash of the HEAD commit.
+ *
+ * @return The hash of the HEAD commit.
+ */
+export async function getHeadHash(): Promise<string> {
+  return (await execGit("rev-parse", ["--verify", "HEAD"])).stdout.toString("utf8");
+}
+
 export async function tagExists(tag: string): Promise<boolean> {
-  const stdout: Buffer = await exec("tag", ["-l", tag]);
-  return stdout.toString("utf8").trim().length > 0;
+  return (await execGit("tag", ["-l", tag])).stdout.toString("utf8").trim().length > 0;
+}
+
+export interface GitCloneOptions {
+  branch?: string;
+  depth?: number;
+  repository: string;
+  directory: AbsPosixPath;
+}
+
+/**
+ * Clone a repository into a new directory
+ */
+export async function gitClone(options: GitCloneOptions): Promise<void> {
+  const args: string[] = [];
+  if (options.branch !== undefined) {
+    args.push("--branch", options.branch);
+  }
+  if (options.depth !== undefined) {
+    args.push("--depth", options.depth.toString(10));
+  }
+  args.push(options.repository, options.directory);
+  await execGit("clone", args);
+}
+
+export interface GitAddOptions {
+  paths: string[];
+  repository: AbsPosixPath;
+}
+
+/**
+ * Clone a repository into a new directory
+ */
+export async function gitAdd(options: GitAddOptions): Promise<void> {
+  const args: string[] = ["--", ...options.paths];
+  await execGit("add", args, {cwd: options.repository});
+}
+
+export interface GitCommitOptions {
+  message: string;
+  author?: string;
+  repository: AbsPosixPath;
+}
+
+export async function gitCommit(options: GitCommitOptions): Promise<void> {
+  const args: string[] = ["-m", options.message];
+  if (options.author !== undefined) {
+    args.push("--author", options.author);
+  }
+  await execGit("commit", args, {cwd: options.repository});
+}
+
+export interface GitPushOptions {
+  local: AbsPosixPath;
+  remote: string;
+}
+
+export async function gitPush(options: GitPushOptions): Promise<void> {
+  await execGit("push", [options.remote], {cwd: options.local});
+  // if (result.stderr.length > 0) {
+  //   throw new Incident("GitPush", result.stderr.toString("utf8"));
+  // }
 }