+++ src/lib/utils/node-async.ts
@@ -1,8 +1,7 @@
-/* tslint:disable:no-angle-bracket-type-assertion */
-
-import * as childProcess from "child_process";
+import {ChildProcess, execFile as _execFile, ExecFileOptions as _ExecFileOptions, spawn as _spawn} from "child_process";
 import * as fs from "fs";
 import {Incident} from "incident";
+import {PassThrough, Transform as TransformStream} from "stream";
 import {promisify} from "util";
 
 export interface ExecFileOptions {
@@ -61,12 +60,6 @@ export class ExecFileError extends Incident<"ExecFileError", ExecFileErrorData, 
   }
 }
 
-export interface NodeAsync {
-  readText(file: string): Promise<string>;
-  writeText(file: string, text: string): Promise<void>;
-  execFile(file: string, args: string[], options?: ExecFileOptions): Promise<ExecFileResult>;
-}
-
 const _readFile: (filename: string, encoding: string) => Promise<string> = <any> promisify(fs.readFile);
 const _writeFile: (filename: string, data: any) => Promise<any> = <any> promisify(fs.writeFile);
 
@@ -80,13 +73,9 @@ export async function writeText(file: string, text: string): Promise<void> {
 
 export async function execFile(file: string, args: string[], options?: ExecFileOptions): Promise<ExecFileResult> {
   return new Promise<ExecFileResult>((resolve, reject) => {
-    const normalizedOptions: childProcess.ExecFileOptions & {encoding: string} = <any> {};
-    if (options !== undefined) {
-      Object.assign(normalizedOptions, options);
-    }
-    normalizedOptions.encoding = "buffer";
+    const normalizedOptions: _ExecFileOptions & {encoding: string} = {...options, encoding: "buffer"};
 
-    childProcess.execFile(
+    _execFile(
       file,
       args,
       normalizedOptions,
@@ -105,4 +94,98 @@ export async function execFile(file: string, args: string[], options?: ExecFileO
   });
 }
 
-export default <NodeAsync> {readText, writeText, execFile};
+export interface SpawnOptions {
+  cwd?: string;
+  env?: {[key: string]: string};
+  stdio: "inherit";
+}
+
+export interface SpawnResult {
+  stdout: Buffer;
+  stderr: Buffer;
+  exit: Exit;
+}
+
+export type Exit = SignalExit | CodeExit;
+
+export interface CodeExit {
+  type: "code";
+  code: number;
+}
+
+export interface SignalExit {
+  type: "signal";
+  signal: string;
+}
+
+export class SpawnedProcess {
+  readonly process: ChildProcess;
+
+  private readonly stdoutChunks: Buffer[];
+  private readonly stderrChunks: Buffer[];
+  private exit?: Exit;
+
+  constructor(file: string, args: string[], options: SpawnOptions) {
+    this.stdoutChunks = [];
+    this.stderrChunks = [];
+    this.exit = undefined;
+
+    this.process = _spawn(
+      file,
+      args,
+      {stdio: [process.stdin, "pipe", "pipe"], cwd: options.cwd, env: options.env},
+    );
+
+    const stdout: TransformStream = new PassThrough();
+    this.process.stdout.pipe(stdout);
+    stdout.pipe(process.stdout);
+    const stderr: TransformStream = new PassThrough();
+    this.process.stderr.pipe(stderr);
+    stderr.pipe(process.stderr);
+
+    stdout.on("data", (chunk: Buffer): void => {
+      this.stdoutChunks.push(chunk);
+    });
+    stderr.on("data", (chunk: Buffer): void => {
+      this.stderrChunks.push(chunk);
+    });
+
+    this.process.once("exit", (code: number | null, signal: string | null): void => {
+      if (code !== null) {
+        this.exit = {type: "code", code};
+      } else {
+        this.exit = {type: "signal", signal: signal!};
+      }
+    });
+  }
+
+  async toPromise(): Promise<SpawnResult> {
+    return new Promise<SpawnResult>((resolve: (res: SpawnResult) => void, reject) => {
+      if (this.exit !== undefined) {
+        const [stdout, stderr]: [Buffer, Buffer] = this.getBuffers();
+        resolve({stdout, stderr, exit: this.exit});
+      } else {
+        this.process.once("exit", (code: number | null, signal: string | null): void => {
+          let exit: Exit;
+          if (code !== null) {
+            exit = {type: "code", code};
+          } else {
+            exit = {type: "signal", signal: signal!};
+          }
+          const [stdout, stderr]: [Buffer, Buffer] = this.getBuffers();
+          resolve({stdout, stderr, exit});
+        });
+      }
+    });
+  }
+
+  private getBuffers(): [Buffer, Buffer] {
+    const stdout: Buffer = Buffer.concat(this.stdoutChunks);
+    const stderr: Buffer = Buffer.concat(this.stderrChunks);
+    this.stdoutChunks.length = 0;
+    this.stderrChunks.length = 0;
+    this.stdoutChunks.push(stdout);
+    this.stderrChunks.push(stderr);
+    return [stdout, stderr];
+  }
+}